+++ test/potFormater.test.js
@@ -24,3 +24,32 @@ it('should return pot formatted string', () => {
     }),
   ).toMatchSnapshot();
 });
+
+it('should return pot formatted string, with null or undefined description', () => {
+  expect(
+    potFormater({
+      'Go to MCS website': [
+        {
+          id: 'NotFound.errorButton',
+          defaultMessage: 'Go to MCS website',
+          filename: './messages/src/containers/NotFound/messages.json',
+        },
+      ],
+    }),
+  ).toMatchSnapshot();
+});
+
+it('should return pot formatted string, with multi line values', () => {
+  expect(
+    potFormater({
+      'NotFound.errorButton': [
+        {
+          id: 'NotFound.errorButton',
+          description: 'My description\nis\nquite\nlong.',
+          defaultMessage: 'This is\nmultiline',
+          filename: './messages/src/containers/NotFound/messages.json',
+        },
+      ],
+    }),
+  ).toMatchSnapshot();
+});